同步@binsee的提交
## 修复问题
- 无法将`转发HTTPS音源地址`设置为空
- 日志页面无法显示完整日志
- 修复防火墙策略
    之前的防火墙策略只能使非ssl连接生效，导致部分平台的客户端使用出现问题
- 修复启动时获取域名列表存入ipset
    之前被人移除了，服务不会立马生效，只有触发网卡连接事件时才更新
- 更新功能中，如果本地不存在证书，则使用默认证书
    适用于本地无证书/本地没有**UnblockNeteaseMusic**包的情况

## 修改/调整
- 脚本启动顺序重新改为99
    原顺序为80，作为一个第三方服务不需要那么靠前的启动顺序，且如ss、koolproxy之类的服务启动顺序为99，他们启动后会修改防火墙策略，有可能导致unblockmusic的防火墙策略顺序靠后导致无法执行到（比如koolproxy）
- 不重启`dnsmasq`服务，改为`重载`，避免影响dns解析

## 优化
- 移除多余无用的配置项
- 增加对`KOOLPROXY`的防火墙策略
   如果`openwrt`上启用了`KOOLPROXY`，且防火墙策略优先于unblockmusic的策略，会导致访问不走unblockmusic，此项对此进行了处理。对于未启用`KOOLPROXY`的无影响，**可忽略**
- 简化启动**UnblockNeteaseMusic**的进程数量。

## 其他
-作者@maxlicheng未测试，若遇问题，欢迎开issues讨论；
-再次感谢[binsee](https://github.com/binsee)

+++ app/root/usr/share/UnblockNeteaseMusic/update_core.sh
@@ -37,7 +37,9 @@ function update_core(){
 
 	wget-ssl --no-check-certificate -t 1 -T 10 -O  /tmp/unblockneteasemusic/core/core.tar.gz "https://github.com/nondanee/UnblockNeteaseMusic/archive/master.tar.gz"  >/dev/null 2>&1
 	tar -zxf "/tmp/unblockneteasemusic/core/core.tar.gz" -C "/tmp/unblockneteasemusic/core/" >/dev/null 2>&1
-	rm -f /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/ca.crt /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/server.crt /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/server.key
+	if [ -e "/usr/share/UnblockNeteaseMusic/ca.crt" ] && [ -e "/usr/share/UnblockNeteaseMusic/server.crt" ] && [ -e "/usr/share/UnblockNeteaseMusic/server.key" ] ; then
+		rm -f /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/ca.crt /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/server.crt /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/server.key
+	fi
 	cp -a /tmp/unblockneteasemusic/core/UnblockNeteaseMusic-master/* "/usr/share/UnblockNeteaseMusic/"
 	rm -rf "/tmp/unblockneteasemusic" >/dev/null 2>&1
